requirements.txt
click==8.1.3
tqdm==4.64.0
faiss-gpu==1.7.2
pillow==9.1.1
numpy==1.21.6
torch==1.11.0
torchvision==0.12.0
scikit-learn==1.0.2
pandas==1.2.4
